Tests/LibWeb/Text/expected/geometry/getClientRects-wrapped-rtl-span-border.txt

rects: 2
wrap[0]: 23,5 32x10
wrap[1]: 17,25 38x10
